requirements.txt
absl-py==0.5.0
APScheduler==3.3.1
astor==0.7.1
astroid==1.6.1
beautifulsoup4==4.6.3
beets==1.4.5
certifi==2017.4.17
chardet==3.0.4
click==6.7
config==0.3.9
cssselect==1.0.3
et-xmlfile==1.0.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.12.1
Flask-Caching==1.4.0
Flask-Cors==3.0.3
Flask-SQLAlchemy==2.3.2
Flask-User==0.6.21
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
idna==2.6
isort==4.3.2
itsdangerous==0.24
jdcal==1.3
jellyfish==0.5.6
jieba3k==0.35.1
Jinja2==2.9.6
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.3.1
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
mccabe==0.6.1
munkres==1.0.12
musicbrainzngs==0.6
mutagen==1.38
mysql-connector-python-rf==2.2.2
MySQL-python==1.2.4
newspaper3k==0.2.8
nltk==3.3
numpy==1.15.2
openpyxl==2.4.8
pandas==0.23.4
Pillow==5.3.0
protobuf==3.6.1
psycopg2==2.7.1
pydub==0.20.0
pylint==1.8.2
PyMySQL==0.9.3
python-dateutil==2.7.3
pytz==2017.2
PyYAML==3.12
requests==2.18.4
requests-file==1.4.3
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.16
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
textblob==0.15.2
tinysegmenter==0.3
tldextract==2.2.0
tzlocal==1.4
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12.1
wrapt==1.10.11
